requirements.txt
discord-py
flask
python-dotenv
supabase
aiohttp
psutil